[Chore] interface 모듈화(#1)
