+++ interface/dinner.ts
@@ -1,0 +1,7 @@
+import { Members } from "./member";
+
+export interface Dinner {
+  member: Members[];
+  shuffle(member: Members[]): Members[];
+  organize(member: Members[]): void;
+}
