+++ interface/member.ts
@@ -1,0 +1,6 @@
+interface Members {
+  name: String;
+  group: String;
+}
+
+export { Members };
